+++ src/main/java/songbox/house/task/SetBitRateTask.java
@@ -6,11 +6,10 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import songbox.house.domain.entity.VkAudio;
 import songbox.house.service.search.vk.VkFileSizeService;
-import songbox.house.util.ThreadLocalAuth;
+import songbox.house.util.ThreadChange;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.Callable;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static java.lang.Float.compare;
@@ -24,7 +23,7 @@ import static songbox.house.util.VkUtil.getPartUrl;
 @Data
 @FieldDefaults(makeFinal = true, level = PRIVATE)
 @Slf4j
-public class SetBitRateTask extends ThreadLocalAuth.LocalAuthCallable<Optional<VkAudio>> {
+public class SetBitRateTask extends ThreadChange.LocalAuthCallable<Optional<VkAudio>> {
 
     VkAudio vkAudio;
     VkFileSizeService vkFileSizeService;
