+++ src/main/java/songbox/house/task/VkTrackDownloader.java
@@ -41,7 +41,7 @@ import static java.util.concurrent.Executors.newFixedThreadPool;
 import static javax.crypto.Cipher.DECRYPT_MODE;
 import static javax.crypto.Cipher.getInstance;
 import static songbox.house.util.Constants.PERFORMANCE_MARKER;
-import static songbox.house.util.ThreadLocalAuth.applyContext;
+import static songbox.house.util.ThreadChange.applyContext;
 import static songbox.house.util.VkUtil.getPartUrl;
 
 @Slf4j
